refactor: make code more straightforward
