+++ packages/bootstrap-vue-next/src/components/BPopover.vue
@@ -75,30 +75,29 @@ import {
   size as sizeMiddleware,
   useFloating,
 } from '@floating-ui/vue'
-import {
-  BvTriggerableEvent,
-  getTransitionDelay,
-  IS_BROWSER,
-  resolveBootstrapPlacement,
-} from '../utils'
-import {DefaultAllowlist, sanitizeHtml} from '../utils/sanitizer'
 import {onClickOutside, useMouseInElement, useToNumber} from '@vueuse/core'
 import {
-  type ComponentPublicInstance,
   computed,
   type CSSProperties,
-  type MaybeRefOrGetter,
   nextTick,
   onBeforeUnmount,
   onMounted,
   ref,
   toRef,
-  toValue,
   watch,
   watchEffect,
 } from 'vue'
 import {useBooleanish, useId} from '../composables'
 import type {BPopoverProps} from '../types'
+import {
+  BvTriggerableEvent,
+  getElement,
+  getTransitionDelay,
+  IS_BROWSER,
+  resolveBootstrapPlacement,
+} from '../utils'
+import {DefaultAllowlist, sanitizeHtml} from '../utils/sanitizer'
+
 defineOptions({
   inheritAttrs: false,
 })
@@ -428,19 +427,6 @@ defineExpose({
   show,
   toggle,
 })
-
-const getElement = (
-  target: MaybeRefOrGetter<string | ComponentPublicInstance | HTMLElement | null>
-): HTMLElement | undefined => {
-  const element = toValue(target)
-  if (!element) return undefined
-  if (typeof element === 'string') {
-    if (typeof document === 'undefined') return undefined
-    const idElement = document.getElementById(element)
-    return idElement ? idElement : (document.querySelector(element) as HTMLElement) || undefined
-  }
-  return (element as ComponentPublicInstance).$el ?? element
-}
 
 const bind = () => {
   // TODO: is this the best way to bind the events?
