+++ packages/bootstrap-vue-next/src/utils/getElement.ts
@@ -1,0 +1,13 @@
+import type {ComponentPublicInstance} from 'vue'
+
+export default (
+  element: string | ComponentPublicInstance | HTMLElement | null
+): HTMLElement | undefined => {
+  if (!element) return undefined
+  if (typeof element === 'string') {
+    if (typeof document === 'undefined') return undefined
+    const idElement = document.getElementById(element)
+    return idElement ?? (document.querySelector(element) as HTMLElement) ?? undefined
+  }
+  return (element as ComponentPublicInstance).$el ?? element
+}
