+++ packages/bootstrap-vue-next/src/utils/index.ts
@@ -6,9 +6,10 @@ export {default as filterEvent} from './filterEvent'
 export * from './floatingUi'
 export {default as getBreakpointProps} from './getBreakpointProps'
 export {default as getClasses} from './getClasses'
-export {default as getTableFieldHeadLabel} from './getTableFieldHeadLabel'
+export {default as getElement} from './getElement'
 export {default as getId} from './getId'
 export {default as getSlotElements} from './getSlotElements'
+export {default as getTableFieldHeadLabel} from './getTableFieldHeadLabel'
 export * from './inspect'
 export {default as isLink} from './isLink'
 export * from './keys'
